fix possible string deallocation issue in path argument of open_file
When Python strings are converted to char * then the lifetime of the C string depends on the Python string. When returning a char * from a function that was created in that function from a Python string then the memory could be reclaimed due to GC and lead to bugs. Returning as a Python string extends the lifetime to the caller who is then responsible for it.